+++ NINA/Utility/Converters/MinusOneToBaseValueConverter.cs
@@ -21,9 +21,13 @@ namespace NINA.Utility.Converters {
     public class MinusOneToBaseValueConverter : IMultiValueConverter {
 
         public object Convert(object[] value, Type targetType, object parameter, CultureInfo culture) {
-            if (int.TryParse(value[0] + "", out var result)) {
-                if (result == -1) {
-                    if(value[1].ToString() == "-1") {
+            var param = new object[] { value[0].ToString(), null };
+            var parsed = (bool)value[0].GetType()
+                .GetMethod("TryParse", new[] { typeof(string), value[0].GetType().MakeByRefType() })
+                .Invoke(null, param);
+            if (parsed) {
+                if (param[1].Equals(System.Convert.ChangeType(-1, param[1].GetType()))) {
+                    if (value[1].ToString() == "-1") {
                         value[1] = Locale.Loc.Instance["LblCamera"];
                     }
                     return "(" + value[1] + ")";
@@ -36,8 +40,12 @@ namespace NINA.Utility.Converters {
         }
 
         public object[] ConvertBack(object value, Type[] targetType, object parameter, CultureInfo culture) {
-            if (int.TryParse(value.ToString(), out int result)) {
-                return new object[] { result };
+            var param = new object[] { value.ToString(), null };
+            var parsed = (bool)targetType[0]
+                .GetMethod("TryParse", new[] { typeof(string), targetType[0].MakeByRefType() })
+                .Invoke(null, param);
+            if (parsed) {
+                return new object[] { param[1] };
             }
 
             return new object[] { -1 };